rough in a split
